+++ components/datetime/src/provider/mod.rs
@@ -49,9 +49,11 @@ const _: () = {
     impl_locations_v1_marker!(Baked);
     impl_metazone_generic_names_long_v1_marker!(Baked);
     impl_metazone_generic_names_short_v1_marker!(Baked);
+    impl_metazone_period_v1_marker!(Baked);
     impl_metazone_specific_names_long_v1_marker!(Baked);
     impl_metazone_specific_names_short_v1_marker!(Baked);
     impl_time_zone_essentials_v1_marker!(Baked);
+    impl_zone_offset_period_v1_marker!(Baked);
 
     impl_weekday_names_v1_marker!(Baked);
     impl_day_period_names_v1_marker!(Baked);
@@ -110,9 +112,11 @@ pub const MARKERS: &[DataMarkerInfo] = &[
     time_zones::LocationsV1Marker::INFO,
     time_zones::MetazoneGenericNamesLongV1Marker::INFO,
     time_zones::MetazoneGenericNamesShortV1Marker::INFO,
+    time_zones::MetazonePeriodV1Marker::INFO,
     time_zones::MetazoneSpecificNamesLongV1Marker::INFO,
     time_zones::MetazoneSpecificNamesShortV1Marker::INFO,
     time_zones::TimeZoneEssentialsV1Marker::INFO,
+    time_zones::ZoneOffsetPeriodV1Marker::INFO,
     neo::WeekdayNamesV1Marker::INFO,
     neo::DayPeriodNamesV1Marker::INFO,
     neo::GluePatternV1Marker::INFO,
