+++ provider/data/datetime/stubdata/mod.rs
@@ -46,6 +46,8 @@ include!("time_zone_essentials_v1_marker.rs.data");
 include!("metazone_generic_names_long_v1_marker.rs.data");
 include!("metazone_generic_names_short_v1_marker.rs.data");
 include!("locations_v1_marker.rs.data");
+include!("metazone_period_v1_marker.rs.data");
+include!("zone_offset_period_v1_marker.rs.data");
 include!("metazone_specific_names_long_v1_marker.rs.data");
 include!("metazone_specific_names_short_v1_marker.rs.data");
 /// Marks a type as a data provider. You can then use macros like
@@ -124,6 +126,8 @@ macro_rules! impl_data_provider {
         impl_metazone_generic_names_long_v1_marker!($provider);
         impl_metazone_generic_names_short_v1_marker!($provider);
         impl_locations_v1_marker!($provider);
+        impl_metazone_period_v1_marker!($provider);
+        impl_zone_offset_period_v1_marker!($provider);
         impl_metazone_specific_names_long_v1_marker!($provider);
         impl_metazone_specific_names_short_v1_marker!($provider);
     };
@@ -182,6 +186,8 @@ macro_rules! impl_any_provider {
                     h if h == <icu::datetime::provider::time_zones::MetazoneGenericNamesLongV1Marker as icu_provider::DataMarker>::INFO.path.hashed() => icu_provider::DataProvider::<icu::datetime::provider::time_zones::MetazoneGenericNamesLongV1Marker>::load(self, req).map(icu_provider::DataResponse::wrap_into_any_response),
                     h if h == <icu::datetime::provider::time_zones::MetazoneGenericNamesShortV1Marker as icu_provider::DataMarker>::INFO.path.hashed() => icu_provider::DataProvider::<icu::datetime::provider::time_zones::MetazoneGenericNamesShortV1Marker>::load(self, req).map(icu_provider::DataResponse::wrap_into_any_response),
                     h if h == <icu::datetime::provider::time_zones::LocationsV1Marker as icu_provider::DataMarker>::INFO.path.hashed() => icu_provider::DataProvider::<icu::datetime::provider::time_zones::LocationsV1Marker>::load(self, req).map(icu_provider::DataResponse::wrap_into_any_response),
+                    h if h == <icu::datetime::provider::time_zones::MetazonePeriodV1Marker as icu_provider::DataMarker>::INFO.path.hashed() => icu_provider::DataProvider::<icu::datetime::provider::time_zones::MetazonePeriodV1Marker>::load(self, req).map(icu_provider::DataResponse::wrap_into_any_response),
+                    h if h == <icu::datetime::provider::time_zones::ZoneOffsetPeriodV1Marker as icu_provider::DataMarker>::INFO.path.hashed() => icu_provider::DataProvider::<icu::datetime::provider::time_zones::ZoneOffsetPeriodV1Marker>::load(self, req).map(icu_provider::DataResponse::wrap_into_any_response),
                     h if h == <icu::datetime::provider::time_zones::MetazoneSpecificNamesLongV1Marker as icu_provider::DataMarker>::INFO.path.hashed() => icu_provider::DataProvider::<icu::datetime::provider::time_zones::MetazoneSpecificNamesLongV1Marker>::load(self, req).map(icu_provider::DataResponse::wrap_into_any_response),
                     h if h == <icu::datetime::provider::time_zones::MetazoneSpecificNamesShortV1Marker as icu_provider::DataMarker>::INFO.path.hashed() => icu_provider::DataProvider::<icu::datetime::provider::time_zones::MetazoneSpecificNamesShortV1Marker>::load(self, req).map(icu_provider::DataResponse::wrap_into_any_response),
                     _ => Err(icu_provider::DataErrorKind::MarkerNotFound.with_req(marker, req)),
