+++ provider/data/timezone/fingerprints.csv
@@ -1,5 +1,3 @@
 time_zone/bcp47_to_iana@1, <singleton>, 7630B, 7609B, 7efd2629d86e0cdc
 time_zone/iana_to_bcp47@3, <singleton>, 9551B, 9509B, ef97860b5f688e6d
-time_zone/metazone_period@1, <singleton>, 11097B, 11009B, dcfa50fb39ebd549
-time_zone/offset_period@1, <singleton>, 13374B, 13286B, 397b18d3f2345411
 time_zone/windows_zones_to_bcp47@1, <singleton>, 8634B, 8591B, 90b7fe3f9cadf5bd
