+++ provider/registry/src/lib.rs
@@ -37,12 +37,14 @@ macro_rules! registry(
             icu::collator::provider::CollationMetadataV1Marker = "collator/meta@1",
             icu::collator::provider::CollationReorderingV1Marker = "collator/reord@1",
             icu::collator::provider::CollationSpecialPrimariesV1Marker = "collator/prim@1",
+            icu::datetime::provider::time_zones::LocationsV1Marker = "time_zone/locations@1",
             icu::datetime::provider::time_zones::MetazoneGenericNamesLongV1Marker = "time_zone/generic_long@1",
             icu::datetime::provider::time_zones::MetazoneGenericNamesShortV1Marker = "time_zone/generic_short@1",
+            icu::datetime::provider::time_zones::MetazonePeriodV1Marker = "time_zone/metazone_period@1",
             icu::datetime::provider::time_zones::MetazoneSpecificNamesLongV1Marker = "time_zone/specific_long@1",
             icu::datetime::provider::time_zones::MetazoneSpecificNamesShortV1Marker = "time_zone/specific_short@1",
             icu::datetime::provider::time_zones::TimeZoneEssentialsV1Marker = "time_zone/essentials@1",
-            icu::datetime::provider::time_zones::LocationsV1Marker = "time_zone/locations@1",
+            icu::datetime::provider::time_zones::ZoneOffsetPeriodV1Marker = "time_zone/offset_period@1",
             icu::decimal::provider::DecimalSymbolsV1Marker = "decimal/symbols@1",
             icu::list::provider::AndListV2Marker = "list/and@2",
             icu::list::provider::OrListV2Marker = "list/or@2",
@@ -193,11 +195,9 @@ macro_rules! registry(
             icu::segmenter::provider::SentenceBreakDataV2Marker = "segmenter/sentence@2",
             icu::segmenter::provider::WordBreakDataOverrideV1Marker = "segmenter/word/override@1",
             icu::segmenter::provider::WordBreakDataV2Marker = "segmenter/word@2",
-            icu::timezone::provider::MetazonePeriodV1Marker = "time_zone/metazone_period@1",
             icu::timezone::provider::names::Bcp47ToIanaMapV1Marker = "time_zone/bcp47_to_iana@1",
             icu::timezone::provider::names::IanaToBcp47MapV3Marker = "time_zone/iana_to_bcp47@3",
             icu::timezone::provider::windows::WindowsZonesToBcp47MapV1Marker = "time_zone/windows_zones_to_bcp47@1",
-            icu::timezone::provider::ZoneOffsetPeriodV1Marker = "time_zone/offset_period@1",
             icu::datetime::provider::neo::WeekdayNamesV1Marker = "datetime/names/weekdays@1",
             icu::datetime::provider::neo::DayPeriodNamesV1Marker = "datetime/names/dayperiods@1",
             icu::datetime::provider::neo::GluePatternV1Marker = "datetime/patterns/glue@1",
